+++ src/main/java/com/cardano/explorer/projection/CollateralInputOutputProjection.java
@@ -1,9 +1,9 @@
 package com.cardano.explorer.projection;
 
-import java.math.BigDecimal;
+import java.math.BigInteger;
 
 public interface CollateralInputOutputProjection {
   String getAddress();
   String getTxHash();
-  BigDecimal getValue();
+  BigInteger getValue();
 }
