+++ src/main/java/com/cardano/explorer/repository/AddressTxBalanceRepository.java
@@ -2,7 +2,7 @@ package com.cardano.explorer.repository;
 
 import com.sotatek.cardano.common.entity.AddressTxBalance;
 import com.sotatek.cardano.common.entity.Tx;
-import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.sql.Timestamp;
 import java.util.List;
 import org.springframework.data.domain.Page;
@@ -20,14 +20,14 @@ public interface AddressTxBalanceRepository extends JpaRepository<AddressTxBalan
       + " INNER JOIN Address addr ON addressTxBalance.address = addr"
       + " WHERE addr.address = :address"
       + " AND addressTxBalance.time < :time")
-  BigDecimal getBalanceByAddressAndTime(String address, Timestamp time);
+  BigInteger getBalanceByAddressAndTime(String address, Timestamp time);
 
   @Query("SELECT addrTxBalance.balance FROM AddressTxBalance addrTxBalance"
       + " INNER JOIN Address addr ON addrTxBalance.address = addr"
       + " INNER JOIN Tx tx ON addrTxBalance.tx = tx"
       + " WHERE addr.address = :address"
       + " ORDER BY addrTxBalance.tx.blockId ASC, addrTxBalance.tx.blockIndex ASC")
-  List<BigDecimal> findAllByAddress(String address);
+  List<BigInteger> findAllByAddress(String address);
 
   @Query(value = "SELECT tx FROM AddressTxBalance addrTxBalance"
       + " INNER JOIN Tx tx ON addrTxBalance.tx = tx"
